pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "inaki"
version = "2.0.0"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "aiosqlite>=0.20",
    "sqlite-vec>=0.1",
    "onnxruntime>=1.18",
    "numpy>=1.26",
    "structlog>=24.0",
    "python-telegram-bot>=21.0",
    "fastapi>=0.111",
    "uvicorn>=0.30",
    "duckduckgo-search>=6.0",
    "croniter>=2.0",
    "exchangelib>=5.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.14",
    "freezegun>=1.5",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
inaki = "main:main"

[tool.setuptools.packages.find]
include = ["core*", "adapters*", "infrastructure*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
